requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
openai==1.2.3
pydantic==2.5.1
pydantic_core==2.14.3
python-docx==1.1.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
gunicorn==20.1.0
